Fix config permissions for Monasca rsyslog plugin
The Monasca rsyslog plugin shares a config folder with the Monasca
Agent. The plugin runs as the syslog user, so the syslog user
must be able to read the config file. This changes adds the syslog
user to a Monasca group therefore granting it access.

An alternative with slightly better security could be to move the
plugin config to a separate directory and give only syslog access
to that. However, the secrets used by the rsyslog plugin are
normally the same as those used by the Monasca Agent.